+++ CHANGES.md
@@ -12,6 +12,7 @@ and this project adheres to
 
 - Add HMAC signing to protect test servers against unauthorized use
 - Add test timeouts (see #Test.Timeout) and exerciser in tests/timeout
+- Add antler init command to create a sample project
 
 ### Changed